migrate import log to mysql table, place mysql params in variables

+++ bin/ftpImport.sh
@@ -5,16 +5,21 @@ DATA_PATH="../downloaded/"
 LOG_PATH="../logs/"
 TMP_PATH="../tmp/"
 
+USER="tomb"
+HOST="184.105.184.30"
+PW="DW4mediatb"
+DB="DWA_SF_Cookie"
+
 # first, retrieve list of available files from server. 
-ftp -pivd ftp.platform.mediamind.com > $LOG_PATH/transfer.log << EOF
+ftp -pid ftp.platform.mediamind.com > $LOG_PATH/transfer.log << EOF
 nlist . $TMP_PATH/oFile
 quit
 EOF
 
 
 # strip out extra formatting
-cat $TMP_PATH/oFile | grep 140102\.zip > $TMP_PATH/files.txt
-rm -f $TMP_PATH/oFile
+cat $TMP_PATH/oFile | grep 140103\.zip > $TMP_PATH/files.txt
+#rm -f $TMP_PATH/oFile
 
 
 rm -f $TMP_PATH/toDownload
@@ -23,17 +28,21 @@ touch $TMP_PATH/toDownload
 rm -f $TMP_PATH/downloadCmds
 touch $TMP_PATH/downloadCmds
 
+rm -f $TMP_PATH/downloaded
+mysql -h$HOST -u$USER -p$PW $DB -e "SELECT fileName FROM import_log" > $TMP_PATH/downloaded
+
 
 while read l; do
 	# if filename is not in .downloaded, add to list of files to retrieve
-	if [ -z `cat .downloaded | grep $l` ] 
+	if [ -z `cat $TMP_PATH/downloaded | grep $l` ] 
 	then
+		echo "retrieving {$l}"
 		echo "get $l $DATA_PATH/$l" >> $TMP_PATH/downloadCmds
 		echo $l >> $TMP_PATH/toDownload
 	fi
 	
 done < $TMP_PATH/files.txt
-
+echo "unzip commands gathered"
 # retrieve files. 
 ftp -vpi ftp.platform.mediamind.com < $TMP_PATH/downloadCmds 2> $LOG_PATH/ftpErrors.log
 
@@ -43,7 +52,7 @@ then
 	exit 
 fi
 
-rm -f $TMP_PATH/downloadCmds
+#rm -f $TMP_PATH/downloadCmds
 cat $TMP_PATH/toDownload >> .downloaded
 
 
@@ -62,15 +71,20 @@ do
 	echo $i
 for f in $DATA_PATH/$i/*.zip; do
 	echo $f
-	unzip -xdu $f >$LOG_PATH/unzip.log
+	unzip -xdu $f >$LOG_PATH/unzip.log 2> $LOG_PATH/unzipErrors.log
 	for gg in u/*.csv; do
 		echo $gg
 		echo "started at:"
 		echo `date`
 		sed -e "s@xxxxCSVFILExxxx@${gg}@g" ../SQL/Load_${i}.sql > tmpSql.sql
-		cat tmpSql.sql | mysql -h184.105.184.30 -utomb -pDW4mediatb --local-infile # ADD ERROR CHECKING
-		rm tmpSql.sql
-		rm $gg
+		cat tmpSql.sql | mysql -h$HOST -u$USER -p$PW $DB --local-infile # ADD ERROR CHECKING
+		if [ "$?" -eq 0 ]
+			then
+				mysql -h$HOST -u$USER -p$PW $DB -e "INSERT IGNORE INTO import_log VALUE ('{$f}', CURRENT_DATE())"
+				rm tmpSql.sql
+				rm $gg
+		fi
+		
 	done
 	
 	mv $f $DATA_PATH/${i}/inSQL
